Some work on rect mesh, basic cubature integration.

+++ common.rs
@@ -1,8 +1,8 @@
 
-pub type R = float;
+pub type R = f64;
 
 #[deriving(Eq, IterBytes, TotalOrd, TotalEq, Clone, Ord)]
 pub struct Deg(u8);
 
 #[deriving(Eq, IterBytes, TotalOrd, TotalEq, Clone, Ord)]
-pub struct Dim(u8);
+pub struct Dim(uint);
